// client/src/app/api/model/raceDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RaceTypeDTO } from './raceTypeDTO';
import { RaceFormatDTO } from './raceFormatDTO';
import { CityDTO } from './cityDTO';


/**
 * Соревнование
 */
export interface RaceDTO { 
    /**
     * Идентификатор
     */
    id: number;
    /**
     * Название
     */
    name: string;
    /**
     * Описание
     */
    description?: string;
    /**
     * Тип
     */
    raceType: RaceTypeDTO;
    /**
     * Дата проведения
     */
    date?: string;
    /**
     * Город проведения
     */
    city?: CityDTO;
    /**
     * Состояние
     */
    state?: string;
    /**
     * Форматы соревнования
     */
    formats?: Array<RaceFormatDTO>;
}